+++ src/types/d2l-types.ts
@@ -150,3 +150,12 @@ export interface FilteredPermutation {
     strength: number[];
   };
 }
+
+export interface FragmentStatModifications {
+  mobility: number;
+  resilience: number;
+  recovery: number;
+  discipline: number;
+  intellect: number;
+  strength: number;
+}
